+++ src/sandbox/my-component.js
@@ -1,4 +1,3 @@
-import React from "react";
 import styled from "styled-components";
 
 const H2 = styled.h2`
